+++ src/firewallfabrik/__init__.py
@@ -10,4 +10,4 @@
 #
 # SPDX-License-Identifier: GPL-2.0-or-later
 
-__version__ = '3.1.8'
+__version__ = '3.1.9'